[API-96] Split api/alerts/index.ts into write/read/ack; index.ts becomes a barrel.
